requirements.txt
accelerate==1.0.1
addict==2.4.0
asttokens==2.4.1
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.0
cycler==0.12.1
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
diffusers==0.31.0
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
Flask==3.0.3
fonttools==4.54.1
fsspec==2024.10.0
idna==3.10
importlib_metadata==8.5.0
ipython==8.28.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiui==0.2.14
kiwisolver==1.4.7
kornia==0.7.4
kornia_rs==0.1.7
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
objprint==0.2.3
opencv-python==4.10.0.84
optional-django==0.1.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.2.0
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
retrying==1.3.4
rpds-py==0.20.0
safetensors==0.4.5
scipy==1.14.1
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
timm==1.0.11
tokenizers==0.20.1
torch==2.5.0
torchvision==0.20.0
tqdm==4.66.5
traitlets==5.14.3
transformers==4.46.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
varname==0.13.5
wcwidth==0.2.13
Werkzeug==3.0.4
widgetsnbextension==4.0.13
xformers==0.0.28.post3
zipp==3.20.2